voice-backend/main_pipecat_official.py
#!/usr/bin/env python3
#
# Official Pipecat-compliant Magentic Voice Backend
# Based on Pipecat examples and best practices
#

import os
import asyncio
import time
from datetime import datetime
from typing import Dict, Any

from dotenv import load_dotenv
from loguru import logger

from pipecat.audio.vad.silero import SileroVADAnalyzer
from pipecat.audio.vad.vad_analyzer import VADParams
from pipecat.frames.frames import LLMRunFrame, TextFrame
from pipecat.pipeline.pipeline import Pipeline
from pipecat.pipeline.runner import PipelineRunner
from pipecat.pipeline.task import PipelineParams, PipelineTask
from pipecat.processors.aggregators.llm_context import LLMContext
from pipecat.processors.aggregators.llm_response import LLMUserAggregatorParams
from pipecat.processors.aggregators.llm_response_universal import LLMContextAggregatorPair
from pipecat.processors.frame_processor import FrameDirection, FrameProcessor
from pipecat.runner.types import RunnerArguments
from pipecat.runner.utils import create_transport
from pipecat.services.groq.llm import GroqLLMService
from pipecat.services.groq.stt import GroqSTTService
from pipecat.services.groq.tts import GroqTTSService
from pipecat.transports.base_transport import BaseTransport, TransportParams
from pipecat.transports.websocket.fastapi import FastAPIWebsocketParams

load_dotenv(override=True)

class MagenticDelegationProcessor(FrameProcessor):
    """Working agent delegation processor - properly integrated"""
    
    def __init__(self):
        super().__init__()
        self.delegation_count = 0
        logger.info("🎯 Working agent delegation processor initialized")
    
    async def process_frame(self, frame, direction: FrameDirection):
        """Process frames with simple, working delegation logic"""
        
        if isinstance(frame, TextFrame):
            text = frame.text
            
            # Simple delegation detection
            delegation_keywords = [
                'create', 'generate', 'make', 'build', 'design',
                'analyze', 'research', 'write', 'develop', 'automate'
            ]
            
            needs_delegation = any(keyword in text.lower() for keyword in delegation_keywords)
            
            if needs_delegation:
                # Simple task classification
                task_type = self._classify_task_type(text)
                
                # Increment counter
                self.delegation_count += 1
                task_id = f"task_{self.delegation_count}_{int(datetime.now().timestamp())}"
                
                # Create simple response
                response_text = f"✅ Task delegated to {task_type} team! Your request '{text}' has been assigned as {task_id}. The team will work on this and provide results through the Magentic-UI interface."
                
                # Log the delegation
                logger.info(f"🎯 Delegated task {task_id}: {text[:50]}... → {task_type} team")
                
                # Create response frame
                response_frame = TextFrame(response_text)
                await self.push_frame(response_frame, direction)
                return
        
        # Pass through other frames
        await self.push_frame(frame, direction)
    
    def _classify_task_type(self, text: str) -> str:
        """Simple task classification"""
        text_lower = text.lower()
        
        if any(word in text_lower for word in ['create', 'generate', 'make', 'build']):
            return 'creation'
        elif any(word in text_lower for word in ['analyze', 'research', 'study']):
            return 'analysis'
        elif any(word in text_lower for word in ['write', 'draft', 'compose']):
            return 'writing'
        elif any(word in text_lower for word in ['design', 'plan', 'architect']):
            return 'design'
        elif any(word in text_lower for word in ['automate', 'script', 'code']):
            return 'automation'
        else:
            return 'general'
    
    def get_status(self) -> Dict[str, Any]:
        """Get delegation status"""
        return {
            "processor_type": "simple_working",
            "total_delegations": self.delegation_count,
            "status": "operational",
            "capabilities": [
                "Task detection and classification",
                "Team assignment",
                "Response generation",
                "Logging and tracking"
            ]
        }

# Transport configuration for different connection types
transport_params = {
    "webrtc": lambda: TransportParams(
        audio_in_enabled=True,
        audio_out_enabled=True,
        vad_analyzer=SileroVADAnalyzer(params=VADParams(stop_secs=0.2)),
    ),
}

async def run_bot(transport: BaseTransport, runner_args: RunnerArguments):
    """Main bot pipeline using official Pipecat structure"""
    logger.info("🚀 Starting Magentic Voice Bot with Pipecat")

    # Initialize Groq services
    stt = GroqSTTService(
        api_key=os.getenv("GROQ_API_KEY"),
        model="whisper-large-v3"
    )

    llm = GroqLLMService(
        api_key=os.getenv("GROQ_API_KEY"),
        model="llama-3.3-70b-versatile"
    )

    tts = GroqTTSService(
        api_key=os.getenv("GROQ_API_KEY"),
        model="playai-tts",
        voice="Fritz-PlayAI"
    )

    # Working agent delegation processor
    delegation_processor = MagenticDelegationProcessor()

    # System messages for the LLM
    messages = [
        {
            "role": "system",
            "content": """You are Frank's personal AI integrator. You help him manage tasks and delegate work to his AutoGen team.

DELEGATION RULES:
- For simple questions, status checks, or casual conversation, respond directly
- For complex tasks that need creation, generation, analysis, or automation, the delegation processor will handle routing to the AutoGen team
- Keep all responses brief and actionable
- Be friendly but professional
- Your output will be converted to audio so don't include special characters in your answers

Examples:
- "How's my team doing?" → Direct response
- "What time is it?" → Direct response  
- "Create a video about our product" → Will be delegated automatically
- "Analyze our sales data" → Will be delegated automatically"""
        },
    ]

    # LLM context and aggregators
    context = LLMContext(messages)
    context_aggregator = LLMContextAggregatorPair(
        context, 
        user_params=LLMUserAggregatorParams(aggregation_timeout=0.05)
    )

    # Create the main pipeline
    pipeline = Pipeline([
        transport.input(),              # Audio/text input from client
        stt,                           # Speech-to-text
        context_aggregator.user(),     # User message aggregation
        delegation_processor,          # AutoGen delegation logic
        llm,                          # Language model processing
        tts,                          # Text-to-speech
        transport.output(),           # Audio/text output to client
        context_aggregator.assistant(), # Assistant response aggregation
    ])

    # Create pipeline task
    task = PipelineTask(
        pipeline,
        params=PipelineParams(
            enable_metrics=True,
            enable_usage_metrics=True,
        ),
        idle_timeout_secs=runner_args.pipeline_idle_timeout_secs,
    )

    # Event handlers
    @transport.event_handler("on_client_connected")
    async def on_client_connected(transport, client):
        logger.info("📱 Client connected to Magentic Voice")
        # Welcome message
        messages.append({
            "role": "system", 
            "content": "Please introduce yourself as Frank's AI integrator and explain that you can help with tasks and delegate complex work to his AutoGen team."
        })
        await task.queue_frames([LLMRunFrame()])

    @transport.event_handler("on_client_disconnected")
    async def on_client_disconnected(transport, client):
        logger.info("📱 Client disconnected from Magentic Voice")
        await task.cancel()

    # Run the pipeline
    runner = PipelineRunner(handle_sigint=runner_args.handle_sigint)
    await runner.run(task)

async def bot(runner_args: RunnerArguments):
    """Main bot entry point compatible with Pipecat Cloud and runners"""
    logger.info("🎤 Initializing Magentic Voice Backend")
    logger.info("🔧 Groq Models: STT=whisper-large-v3, LLM=llama-3.3-70b-versatile, TTS=playai-tts")
    logger.info("🤖 AutoGen integration enabled for task delegation")
    
    # Create transport using Pipecat's official transport creation
    transport = await create_transport(runner_args, transport_params)
    
    # Run the bot
    await run_bot(transport, runner_args)

if __name__ == "__main__":
    # Use Pipecat's official runner
    from pipecat.runner.run import main
    
    # Check for required environment variables
    if not os.getenv("GROQ_API_KEY"):
        logger.error("❌ GROQ_API_KEY environment variable is required")
        exit(1)
    
    logger.info("🎤 Starting Magentic Voice Backend with official Pipecat structure")
    logger.info("📱 Ready for web and mobile clients")
    logger.info("🔧 AutoGen team URL: http://magentic-ui:8081")
    
    # Run with official Pipecat main
    main()
