voice-backend/requirements.txt
# Pipecat framework with all necessary extensions
pipecat-ai[groq,websocket,silero,runner,webrtc]

# Environment and logging
python-dotenv>=1.0.0
loguru>=0.7.0

# HTTP client for AutoGen team communication
httpx>=0.25.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Audio processing
numpy

# WebRTC support
aiortc

# Ultra low-latency optimizations
uvloop>=0.19.0
orjson>=3.9.0
aiohttp>=3.9.0
